# internal/server/openapi.yaml
openapi: 3.0.3
info:
  title: VulnScanner REST API
  version: 1.0.0
  description: |
    VulnScanner 资产漏洞扫描与管理平台 REST API。
    - 鉴权：请求头 `X-API-Key`（admin 级自动化凭证）或用户登录返回的 `Authorization: Bearer <token>`（按角色鉴权），二者任选其一。
    - 角色：`admin` 全部操作；`operator` 日常运维写操作；`viewer` 只读（`/audit-logs*` 仅 admin 可见）。
    - 公开端点：`/health`、`/demo`、`/dl/*`、`/r/*`、`/api/v1/register`、`/api/v1/auth/login`、`/api/v1/auth/ldap/login`、`/openapi.yaml`。
    - 错误格式统一为 `{"error": "<message>"}`。
servers:
  - url: /
...
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /dl/agent/{platform}:
    get:
      tags: [downloads]
      summary: 下载 Agent 安装包
      security: []
      parameters:
        - name: platform
          in: path
          required: true
          schema: {type: string, enum: [linux-amd64, linux-arm64, windows-amd64, windows-arm64]}
      responses:
        '200':
          description: Agent 可执行文件
          content:
            application/octet-stream:
              schema: {type: string, format: binary}
  /dl/{platform}:
    get:
      tags: [downloads]
      summary: 下载 Agent 安装包（兼容别名）
      security: []
      parameters:
        - name: platform
          in: path
          required: true
          schema: {type: string}
      responses:
        '200':
          description: Agent 可执行文件
          content:
            application/octet-stream:
              schema: {type: string, format: binary}
  /r/{code}:
    get:
      tags: [downloads]
      summary: 获取一次性安装脚本
      security: []
      parameters:
        - name: code
          in: path
          required: true
          schema: {type: string}
          description: Agent 注册码
      responses:
        '200':
          description: bash 或 PowerShell 安装脚本
          content:
            text/plain:
              schema: {type: string}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/v1/register:
    post:
      tags: [auth]
      summary: Agent 注册（使用一次性注册码）
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code: {type: string}
                hostname: {type: string}
                os_type: {type: string}
                os_version: {type: string}
                arch: {type: string}
      responses:
        '200':
          description: 注册成功，返回 agent_id 与 gRPC 地址
          content:
            application/json:
              schema:
                type: object
                properties:
                  agent_id: {type: string}
                  grpc_addr: {type: string}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/v1/auth/login:
    post:
      tags: [auth]
      summary: 用户登录
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功，返回 JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401': {$ref: '#/components/responses/Unauthorized'}
  /api/v1/auth/ldap/login:
    post:
      tags: [auth]
      summary: LDAP 目录登录（绑定认证 + 组角色映射 + 自动建号）
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功，返回 JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /api/v1/auth/me:
    get:
      tags: [auth]
      summary: 当前登录用户信息
      x-roles: [admin, operator, viewer]
      responses:
        '200':
          description: 当前用户
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401': {$ref: '#/components/responses/Unauthorized'}
  /api/v1/auth/change-password:
    post:
      tags: [auth]
      summary: 修改本人密码
      x-roles: [admin, operator, viewer]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: 密码已更新
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: {type: string}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /api/v1/tenants:
    get:
      tags: [auth]
      summary: 租户列表（仅系统管理员）
      x-roles: [admin]
      responses:
        '200':
          description: 租户列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  tenants:
                    type: array
                    items:
                      $ref: '#/components/schemas/Tenant'
    post:
      tags: [auth]
      summary: 创建租户（仅系统管理员）
      x-roles: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TenantInput'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'
        '400': {$ref: '#/components/responses/BadRequest'}
  /api/v1/users/{userId}/tenant:
    put:
      tags: [auth]
      summary: 调整用户所属租户（仅系统管理员）
      x-roles: [admin]
      parameters:
        - name: userId
          in: path
          required: true
          schema: {type: integer, format: int64}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenant_id]
              properties:
                tenant_id: {type: integer, format: int64}
      responses:
        '200':
          description: 调整成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id: {type: integer, format: int64}
                  tenant_id: {type: integer, format: int64}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/v1/agents/{agentId}/tenant:
    put:
      tags: [agents]
      summary: 调整 Agent 所属租户（仅系统管理员；资产数据随 Agent 迁移）
      x-roles: [admin]
      parameters:
        - name: agentId
          in: path
          required: true
          schema: {type: string}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenant_id]
              properties:
                tenant_id: {type: integer, format: int64}
      responses:
        '200':
          description: 调整成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  agent_id: {type: string}
                  tenant_id: {type: integer, format: int64}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/v1/users:
    get:
      tags: [auth]
      summary: 用户列表
      x-roles: [admin, operator, viewer]
      responses:
        '200':
          description: 用户列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
    post:
      tags: [auth]
      summary: 创建用户
      x-roles: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400': {$ref: '#/components/responses/BadRequest'}
  /api/v1/users/{userId}:
    put:
      tags: [auth]
      summary: 更新用户（角色/状态/显示名）
      x-roles: [admin]
      parameters:
        - name: userId
          in: path
          required: true
          schema: {type: integer, format: int64}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id: {type: integer, format: int64}
                  status: {type: string}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
    delete:
      tags: [auth]
      summary: 删除用户
      x-roles: [admin]
      parameters:
        - name: userId
          in: path
          required: true
          schema: {type: integer, format: int64}
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id: {type: integer, format: int64}
                  status: {type: string}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/v1/users/{userId}/password:
    post:
      tags: [auth]
      summary: 重置用户密码
      x-roles: [admin]
      parameters:
        - name: userId
          in: path
          required: true
          schema: {type: integer, format: int64}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: 密码已重置
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id: {type: integer, format: int64}
                  status: {type: string}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
...
  /api/v1/agents/{id}/vulns/{cveId}:
    get:
      tags: [vulnerabilities]
      summary: 单条漏洞详情
      x-roles: [admin, operator, viewer]
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
        - {name: cveId, in: path, required: true, schema: {type: string}}
      responses:
        '200':
          description: 漏洞详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CVEResult'
        '404': {$ref: '#/components/responses/NotFound'}
...
  /api/v1/cmdb/reconcile-report:
    get:
      tags: [assets]
      summary: CMDB 对账报告
      x-roles: [admin, operator, viewer]
      responses:
        '200':
          description: 对账报告
          content:
            application/json:
              schema:
                type: object
        '500': {$ref: '#/components/responses/InternalError'}
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: admin 级自动化凭证（等同 admin 角色）。
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 用户登录返回的 JWT，按用户角色鉴权。
  responses:
    BadRequest:
      description: 请求参数或请求体不合法
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: 缺少或无效的鉴权凭证
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    LoginResponse:
      type: object
      properties:
        token: {type: string}
        expires_at: {type: string, format: date-time}
        user:
          $ref: '#/components/schemas/User'
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: {type: string}
        password: {type: string}
...